Добавил фрагменты и функции для просмотра персон, фактов, трейлеров и отзывов.

+++ app/src/main/java/com/example/kplist/data/models/apiReviewModel/Doc.kt
@@ -14,5 +14,5 @@ data class Doc(
     val title: String,
     val type: String,
     val updatedAt: String,
-    val userRating: Int
+    val userRating: Double
 )
